[LFC][IFC] Expand "empty line" fast path support for multiline inline boxes https://bugs.webkit.org/show_bug.cgi?id=221024
Reviewed by Antti Koivisto.

Let's check if the inline box starts at this empty line and update its geometry accordingly.
Note that just because an inline box end (</span>) is on an empty line, it does not necessarily make the inline box
taller (e.g. <div>some content<span><br></span> <- while the </span> is after the <br> on a new and empty line, it does
not stretch the associated inline box).

* layout/inlineformatting/InlineFormattingContext.cpp:
(WebCore::Layout::InlineFormattingContext::computeGeometryForLineContent):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271951 268f45cc-cd09-0410-ab3c-d52691b4dbfc