[GTK] run-gtk-tests: Crashing and non-existent tests should not count as a pass https://bugs.webkit.org/show_bug.cgi?id=220863
Reviewed by Michael Catanzaro.

The current implementation doesn't account for test binaries that have
not emitted any subtests. This is the case when a test binary doesn't
exist or it crashes.

Also, in the latter case, the stderr of the crashed processes was not
being outputted to the user, masking a crashing test binary as a
passing test.

This patch fixes both issues: It adds several warning prints when
binaries can't be run or fail to add any subtest, emitting the stderr
generated by the failing or crashing binary.

* glib/api_test_runner.py:
(TestRunner._run_test):
(TestRunner.run_tests):
* glib/glib_test_runner.py:
(GLibTestRunner.run):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271955 268f45cc-cd09-0410-ab3c-d52691b4dbfc