[MacOS] Enable WebKitTestRunner audio and video capture in GPUProcess https://bugs.webkit.org/show_bug.cgi?id=221035
Reviewed by Eric Carlson.

Tools:

Remove the default TestOptions values and instead on MacOS rely on capture in GPUProcess.
On iOS, capture ill still happen inprocess for the time being.

* WebKitTestRunner/TestOptions.cpp:
(WTR::TestOptions::defaults):

LayoutTests:

fast/mediastream/media-stream-track-interrupted.html needs to be run with mock audio capture in process
as interruption mock testing is only supported this way.
SpeechRecognition in GPUProcess is not yet working so we disable capture in GPU Process for these two tests
until SpeechRecognition in GPUProcess is supported.

* fast/mediastream/media-stream-track-interrupted.html:
* fast/speechrecognition/start-recognition-then-stop.html:
* fast/speechrecognition/start-second-recognition.html:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@272053 268f45cc-cd09-0410-ab3c-d52691b4dbfc