REGRESSION: Fullscreen HTML5 video subtitles/captions are shown twice https://bugs.webkit.org/show_bug.cgi?id=214634
Reviewed by Eric Carlson.

When we need to hide a text track representation, we can hide it directly. But, when we
need to show(set "hidden" to false) a text track representation, we have to render it first.
Otherwise, the previously rendered content will flash for a short period of time.

Covered by manual test.

* html/shadow/MediaControlTextTrackContainerElement.cpp:
(WebCore::MediaControlTextTrackContainerElement::updateDisplay):
(WebCore::MediaControlTextTrackContainerElement::updateTextTrackRepresentationImageIfNeeded):
(WebCore::MediaControlTextTrackContainerElement::updateTextTrackRepresentationIfNeeded):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@272190 268f45cc-cd09-0410-ab3c-d52691b4dbfc