Don't launch a network process when registering URL schemes https://bugs.webkit.org/show_bug.cgi?id=221427
Reviewed by Chris Dumez.

Source/WebKit:

When Safari on macOS is launching, it calls registerURLSchemeAsSecure and registerURLSchemeAsBypassingContentSecurityPolicy.
This causes the network process to launch before the app has finished launching.  Since values are stored in the UI process
and sent in NetworkProcessCreationParameters whenever a network process starts, so it's not necessary to start the network
process when calling these two functions.  This should decrease Safari's launch time.

* UIProcess/WebProcessPool.cpp:
(WebKit::WebProcessPool::registerURLSchemeAsSecure):
(WebKit::WebProcessPool::registerURLSchemeAsBypassingContentSecurityPolicy):

Tools:

* TestWebKitAPI/Tests/WebKitCocoa/NetworkProcess.mm:
(TEST):



git-svn-id: http://svn.webkit.org/repository/webkit/trunk@272383 268f45cc-cd09-0410-ab3c-d52691b4dbfc